+++ selfdrive/test/helpers.py
@@ -14,7 +14,7 @@ from openpilot.system.version import training_version, terms_version
 
 
 def set_params_enabled():
-  os.environ['FINGERPRINT'] = "TOYOTA COROLLA TSS2 2019"
+  os.environ['FINGERPRINT'] = "TOYOTA_COROLLA_TSS2"
   os.environ['LOGPRINT'] = "debug"
 
   params = Params()
@@ -88,23 +88,37 @@ def read_segment_list(segment_list_path):
   return [(platform[2:], segment) for platform, segment in zip(seg_list[::2], seg_list[1::2], strict=True)]
 
 
+@contextlib.contextmanager
+def http_server_context(handler, setup=None):
+  host = '127.0.0.1'
+  server = http.server.HTTPServer((host, 0), handler)
+  port = server.server_port
+  t = threading.Thread(target=server.serve_forever)
+  t.start()
+
+  if setup is not None:
+    setup(host, port)
+
+  try:
+    yield (host, port)
+  finally:
+    server.shutdown()
+    server.server_close()
+    t.join()
+
+
 def with_http_server(func, handler=http.server.BaseHTTPRequestHandler, setup=None):
   @wraps(func)
   def inner(*args, **kwargs):
-    host = '127.0.0.1'
-    server = http.server.HTTPServer((host, 0), handler)
-    port = server.server_port
-    t = threading.Thread(target=server.serve_forever)
-    t.start()
-
-    if setup is not None:
-      setup(host, port)
-
-    try:
-      return func(*args, f'http://{host}:{port}', **kwargs)
-    finally:
-      server.shutdown()
-      server.server_close()
-      t.join()
-
+    with http_server_context(handler, setup) as (host, port):
+      return func(*args, f"http://{host}:{port}", **kwargs)
   return inner
+
+
+def DirectoryHttpServer(directory) -> type[http.server.SimpleHTTPRequestHandler]:
+  # creates an http server that serves files from directory
+  class Handler(http.server.SimpleHTTPRequestHandler):
+    def __init__(self, *args, **kwargs):
+      super().__init__(*args, directory=str(directory), **kwargs)
+
+  return Handler